More accurate beam model and faster archive calls.
I've implemented the analytical beam model thanks to David Kaplan. The Galaxy MWAops group know has admin access to metadata which makes downloading the metadata faster thanks to Dave Pallot. The file containing the password for the admin access will likely not be stored on GitHub as it can be viewed by the public. I've made the output more useful for Mengyao's autofold script.